Fix problem filter
